sha1: Fix type conversion warning

+++ src/brickd/sha1.c
@@ -98,7 +98,7 @@ void sha1_update(SHA1 *sha1, const uint8_t *data, size_t length) {
 	size_t i, j;
 
 	j = (size_t)((sha1->count >> 3) & 63);
-	sha1->count += (length << 3);
+	sha1->count += (uint64_t)length << 3;
 
 	if ((j + length) > 63) {
 		i = 64 - j;
